requirements.txt
streamlit>=1.35,<1.41
openai>=1.30.0,<2.0.0
# Pin LangChain to pre-0.2 to keep `langchain.chains` available
langchain<0.2.0
langchain-openai<0.1.0
langchain-community<0.1.0
faiss-cpu>=1.7.4
PyPDF2==3.0.1
tiktoken>=0.7.0,<0.8.0
requests>=2.31.0,<3.0.0
fastembed>=0.2.0,<0.4.0
python-dotenv>=1.0.0,<2.0.0
passlib>=1.7.4,<2.0.0
pinecone>=3.0.0,<4.0.0
mysql-connector-python>=8.0.0,<9.0.0
streamlit-webrtc>=0.47.0
streamlit-js-eval>=0.1.5
pydub>=0.25.1
SpeechRecognition>=3.10.0
